# Documentation/Backend/API Documentations/swagger_documentation.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Ema2a",
    "version": "v1"
  },
  "paths": {
...
    "/api/SignLanguageTranslator/finalize-sentence": {
      "post": {
        "tags": [
          "SignLanguageTranslator"
        ],
        "parameters": [
          {
            "name": "client",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "frontend"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SentenceData"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SentenceData"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SentenceData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StringAPIResponseDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringAPIResponseDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringAPIResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/SignLanguageTranslator/generate-audio": {
      "post": {
        "tags": [
          "SignLanguageTranslator"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TTSRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TTSRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TTSRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TTSResponseAPIResponseDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TTSResponseAPIResponseDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TTSResponseAPIResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/SignLanguageTranslator/text-to-speech/hardware": {
      "post": {
        "tags": [
          "SignLanguageTranslator"
        ],
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "mp3"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TTSRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TTSRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TTSRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HttpStatusCode": {
        "enum": [
          100,
          101,
          102,
          103,
          200,
          201,
          202,
          203,
          204,
          205,
          206,
          207,
          208,
          226,
          300,
          301,
          302,
          303,
          304,
          305,
          306,
          307,
          308,
          400,
          401,
          402,
          403,
          404,
          405,
          406,
          407,
          408,
          409,
          410,
          411,
          412,
          413,
          414,
          415,
          416,
          417,
          421,
          422,
          423,
          424,
          426,
          428,
          429,
          431,
          451,
          500,
          501,
          502,
          503,
          504,
          505,
          506,
          507,
          508,
          510,
          511
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "SentenceData": {
        "type": "object",
        "properties": {
          "sentence": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StringAPIResponseDTO": {
        "type": "object",
        "properties": {
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean"
          },
          "statusCode": {
            "$ref": "#/components/schemas/HttpStatusCode"
          }
        },
        "additionalProperties": false
      },
...
      "TTSRequest": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TTSResponse": {
        "type": "object",
        "properties": {
          "audioData": {
            "type": "string",
            "nullable": true
          },
          "sampleRate": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "TTSResponseAPIResponseDTO": {
        "type": "object",
        "properties": {
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/TTSResponse"
          },
          "success": {
            "type": "boolean"
          },
          "statusCode": {
            "$ref": "#/components/schemas/HttpStatusCode"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Enter: Bearer {your JWT token}",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}